Moving Unit/Integration Tests to new namespace and structure
