+++ .gitignore
@@ -15,4 +15,6 @@ publish/
 **/TestResults/
 /AAEmu.Game/Properties/PublishProfiles
 /AAEmu.Game/ClientData/*
+/AAEmu.Game/Config.json
+/AAEmu.IntegrationTests/Config.json
 *.sqlite3
