+++ AAEmu.Commons/Exceptions/GameException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace AAEmu.Commons.Exceptions;
+public class GameException : Exception
+{
+    public GameException(string message) : base(message)
+    {
+    }
+}
