+++ AAEmu.Commons/Exceptions/MarshalException.cs
@@ -1,0 +1,9 @@
+﻿namespace AAEmu.Commons.Exceptions
+{
+    public sealed class MarshalException : GameException // next: is it necessary?
+    {
+        public MarshalException() : base("Marshal exception")
+        {
+        }
+    }
+}
