+++ AAEmu.Commons/Utils/StringExtensions.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 
 namespace AAEmu.Commons.Utils
 {
@@ -8,7 +8,8 @@ namespace AAEmu.Commons.Utils
         {
             if (string.IsNullOrEmpty(input))
                 throw new ArgumentNullException($"{nameof(input)} is null or empty");
-            return input[0].ToString().ToUpper() + input.Substring(1);
+
+            return char.ToUpper(input[0]) + input[1..];
         }
     }
 }
