+++ AAEmu.Game/AAEmu.Game.csproj
@@ -11,16 +11,20 @@
 
     <ItemGroup>
         <ProjectReference Include="..\AAEmu.Commons\AAEmu.Commons.csproj" />
-        <InternalsVisibleTo Include="AAEmu.Tests.Integration" />
+        <InternalsVisibleTo Include="AAEmu.IntegrationTests" />
     </ItemGroup>
 
     <ItemGroup>
         <Content Include="Configurations\AccessLevels.json">
           <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </Content>
+        <Content Include="Config.json">
+          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
         <Content Include="NLog.config">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </Content>
+        <None Remove="Config.json" />
         <None Remove="ExampleConfig.xml" />
         <Content Include="ExampleConfig.json">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
