+++ AAEmu.Game/Core/Managers/AiGeodataManager.cs
@@ -15,6 +15,8 @@ using NLog;
 
 using Point = AAEmu.Game.Models.Game.AI.AStar.Point;
 
+#pragma warning disable IDE0048 // Add parentheses for clarity
+
 namespace AAEmu.Game.Core.Managers
 {
     // GeoData AiNavigation
