+++ AAEmu.Game/Core/Packets/C2G/CSCancelLeaveWorldPacket.cs
@@ -12,7 +12,7 @@ namespace AAEmu.Game.Core.Packets.C2G
         {
         }
 
-        public override async void Read(PacketStream stream)
+        public override void Read(PacketStream stream)
         {
             if (Connection?.LeaveTask == null)
                 return;
