+++ AAEmu.IntegrationTests/readme.md
@@ -1,0 +1,6 @@
+# Setting up Integration Tests
+
+1. Create a copy of ExampleConfig.json and save as Config.json
+2. Edit the Config.json %wildcards% using the same configurations you did in AAEmu.Game
+3. Copy your sqlite database file from AAEmu.Game Data to the IntegrationTests Data folder
+4. All ready to run integration tests
