--- AAEmu.Tests.Integration/readme.md
@@ -1,5 +1,0 @@
-# Setting up Integration Tests
-
-1. Create a copy of ExampleConfig.json and save as Config.json
-2. Edit the Config.json %wildcards% using the same configurations you did in AAEmu.Game
-3. All ready to run integration tests
