+++ AAEmu.UnitTests/Game/Core/Managers/World/ExampleTests.cs
@@ -1,0 +1,15 @@
+﻿using AAEmu.Game.Core.Managers.World;
+using Xunit;
+
+namespace AAEmu.UnitTests.Game.Core.Managers.World
+{
+    public class ExampleTests
+    {
+
+        [Fact]
+        public void SampleTest()
+        {
+            Assert.Equal(1, SpecialtyManager.Instance.GetValueOfOne());
+        }
+    }
+}
